Renamed `SimplePersistenceStore` to `ZooKeeperPersistenceStore`
